Refactor Controls in a new module
